move ConversationTracker to its own module
